#41267/ユーザー情報/adminユーザも削除が出来てしまう

+++ SRC/libAdminUser.php
@@ -43,7 +43,10 @@ function subAdminUser()
 						<td class="list_td<?php print $i; ?>"></td>
 						<td class="list_td<?php print $i; ?>"><?php print fnAuthorityName($authority); ?></td>
 						<td class="list_td<?php print $i; ?>">
-							<a href="javascript:fnAdminUserDeleteCheck(<?php print $userNo; ?>,'<?php print $name; ?>');">削除</a>
+							<?php if ($userNo > 1) { ?>
+								<a href="javascript:fnAdminUserDeleteCheck(<?php print $userNo; ?>,'<?php print $name; ?>');">削除</a>
+							<?php } ?>
+						</td>
 						</td>
 					</tr>
 				<?php
